requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0 
anyio==4.13.0 
attrs==26.1.0 
blinker==1.9.0 
cachetools==7.0.5 
certifi==2026.2.25 
charset-normalizer==3.4.7 
click==8.3.2 
colorama==0.4.6 
fastapi==0.135.3 
gitdb==4.0.12 
GitPython==3.1.46 
h11==0.16.0 
idna==3.11 
Jinja2==3.1.6 
joblib==1.5.3 
jsonschema==4.26.0 
jsonschema-specifications==2025.9.1 
MarkupSafe==3.0.3 
narwhals==2.19.0 
numpy==2.4.4 
packaging==26.0 
pandas==3.0.2 
pillow==12.2.0 
protobuf==7.34.1 
pyarrow==23.0.1 
pydantic==2.12.5 
pydantic_core==2.41.5 
pydeck==0.9.1 
python-dateutil==2.9.0.post0 
referencing==0.37.0 
requests==2.33.1 
rpds-py==0.30.0 
scikit-learn==1.8.0 
scipy==1.17.1 
six==1.17.0 
smmap==5.0.3 
starlette==1.0.0 
tenacity==9.1.4 
threadpoolctl==3.6.0 
toml==0.10.2 
tornado==6.5.5 
typing-inspection==0.4.2 
typing_extensions==4.15.0 
tzdata==2026.1 
urllib3==2.6.3 
uvicorn==0.44.0 
watchdog==6.0.0 
streamlit
pandas
numpy
requests
python-dotenv
Pillow
PyJWT
plotly
shap
scikit-learn
streamlit-oauth
matplotlib